# tx-backend/openapi/traceability-foss-backend.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Trace-FOSS - OpenAPI Documentation",
    "description": "Trace-FOSS is a system for tracking parts along the supply chain. A high level of transparency across the supplier network enables faster intervention based on a recorded event in the supply chain. This saves costs by seamlessly tracking parts and creates trust through clearly defined and secure data access by the companies and persons involved in the process.",
    "license": {
      "name": "License: Apache 2.0"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:9999/api",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/bpn-config": {
      "get": {
        "tags": [
          "BpnEdcMapping"
        ],
        "summary": "Get BPN EDC URL mappings",
        "description": "The endpoint returns a result of BPN EDC URL mappings.",
        "operationId": "getBpnEdcs",
        "responses": {
          "401": {
            "description": "Authorization failed.",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BpnEdcMapping"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden.",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BpnEdcMapping"
                  }
                }
              }
            }
          },
          "200": {
            "description": "Returns the paged result found",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BpnEdcMapping"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "oAuth2": [
              "profile email"
            ]
          }
        ]
      },
      "put": {
        "tags": [
          "BpnEdcMapping"
        ],
        "summary": "Updates BPN EDC URL mappings",
        "description": "The endpoint updates BPN EDC URL mappings",
        "operationId": "updateBpnEdcMappings",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/BpnEdcMappingRequest"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Created.",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BpnEdcMapping"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authorization failed.",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BpnEdcMapping"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden.",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BpnEdcMapping"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "oAuth2": [
              "profile email"
            ]
          }
        ]
      },
      "post": {
        "tags": [
          "BpnEdcMapping"
        ],
        "summary": "Creates BPN EDC URL mappings",
        "description": "The endpoint creates BPN EDC URL mappings",
        "operationId": "createBpnEdcUrlMappings",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/BpnEdcMappingRequest"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Created.",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BpnEdcMapping"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authorization failed.",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BpnEdcMapping"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden.",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BpnEdcMapping"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "oAuth2": [
              "profile email"
            ]
          }
        ]
      }
    },
...
    "/bpn-config/{bpn}": {
      "delete": {
        "tags": [
          "BpnEdcMapping"
        ],
        "summary": "Deletes BPN EDC URL mappings",
        "description": "The endpoint deletes BPN EDC URL mappings",
        "operationId": "deleteBpnEdcUrlMappings",
        "parameters": [
          {
            "name": "bpn",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Authorization failed."
          },
          "204": {
            "description": "Deleted."
          },
          "403": {
            "description": "Forbidden."
          }
        },
        "security": [
          {
            "oAuth2": [
              "profile email"
            ]
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "BpnEdcMappingRequest": {
        "required": [
          "bpn",
          "url"
        ],
        "type": "object",
        "properties": {
          "bpn": {
            "type": "string"
          },
          "url": {
            "type": "string"
          }
        }
      },
      "BpnEdcMapping": {
        "type": "object",
        "properties": {
          "bpn": {
            "type": "string"
          },
          "url": {
            "type": "string"
          }
        }
      },
...
    },
...
  }
}